ml_services/requirements.txt
accelerate==1.10.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
diffusers==0.36.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
jinja2==3.1.6
kiwisolver==1.4.7
markupsafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
psutil==7.2.2
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
six==1.17.0
starlette==0.49.3
sympy==1.14.0
timm==1.0.24
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==4.57.6
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zipp==3.23.0
